frontend/pyproject.toml
[project]
name = "frontend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests==2.32.3",
    "streamlit==1.44.1",
    "openai==1.75.0",
    "pandas==2.2.3",
    "python-dotenv==1.0.1",
    "pydantic==2.11.2",
    "torch>=2.2.0",
    "transformers==4.47.1",
    "peft>=0.10.0",
    "numpy<2.0.0",
    "langchain==0.3.13",
    "langchain-openai==0.2.14",
    "langchain-community==0.3.13",
    "duckduckgo-search==7.0.1",
    "uuid>=1.30",
    "PyYAML>=6.0",
]
